gitignore --save-temps products
If you use --save-temps (C Flags) then this will make git ignore the generated files.
The --save-temps compiler flag is useful for debugging preprocessor issues (#include, #define, etc.) but it puts files in the project directory. It's supposed to be able to take a path as an option but it's ignored? And it somehow removes all the source code indexing which makes "Jump to Definition" not work so it shouldn't be set all the time.