Add iofdebug
This is a catch-all for IOFramebuffer debugging and features. It uses vtable patching so it affects all IOFramebuffer drivers (Intel, AMD, Nvidia).
It uses the IOI2CInterface APIs to allow user space apps to do things such as set or get attributes and settings (can set 6bpc color for Nvidia GPUs!) or debug settings such as logging verbosity. This is simpler than setting up an IOUserClient because we're already patching doi2cRequest.

IOFramebuffer_vtable.hpp contains the vtable info. The first argument to onevtableitem is 0 or 1 - set to 1 to enable the patch for that item.